==> RMarkdowns/03_Variograms&Performance_Maps.Rmd <==
---
title: "Appendix 2: Home Range Analysis for the 60 Sampling Regimes"
date: \today
output:
  rmarkdown::pdf_document:
    latex_engine: xelatex
    extra_dependencies: ["float"]
    fig_caption: yes        
    includes:  
      in_header: preamble-latex.txt
header-includes:
  \floatplacement{figure}{H}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "H", out.extra = "")
```

### Background

The purpose of this document is to show the full results of the home range analysis from the 60 sampling regimes (i.e. the results of applying the analytical workflow presented in Appendix 1 to our study). This includes the: 1) variogram regressions with the fitted movement models, 2) the movement model summary information, and 3) home range estimates (See Appendix 1 for a detailed description of how home range analysis in *ctmm* works).

As a reminder, the 60 sampling regimes were generated from portions of our longitudinal movement dataset that were of the highest quality (two or more months of consistent data with very few temporal gaps), which we call "complete segments". There are six complete segments from five groups because two of the complete segments came from the same group at different time periods (AA and AA2). How the sampling regimes were generated from the complete segments is described in the Methods section of the main text. We include the important information for every complete segment (i.e. labelled as "all") and every sampling regime. The names of the sampling regimes are labelled with (C) for concentrated or (S) for spread, followed by the number of days in the data (e.g. C10).

```{r, message=FALSE, warning=FALSE, include = FALSE}
# load packages
library(tidyverse)
library(ctmm)
library(ctmmweb)
library(sf)
library(lubridate)
library(janitor)
```

**Note on variograms:**  These plots are used as visualization tools to see if the data are sufficient to show range residency. Generally, what we are looking for is whether the empirical variograms plateau indicating range residency. If the variograms continue to increase, then this is an immediate sign that the data is not suitable for home range estimation. With irregular data, the variograms can look quite messy, which is generally not a problem. Also, they tend to be a bit more unpredictable toward larger time-lags which also is not normally an issue. In these plots, we also include the mean and confidence intervals for the top movement model (determined by AICc) fitted to the empirical variogram. 

**Note on model summaries:** For the complete segments and sampling regimes from each group, we show a table showing the name of the selected movement model, the home range crossing time and the effective sample size. Overall, the *Ornstein--Uhlenbeck Foraging* (OUF) model was always selected, meaning that the data showed autocorrelated positions and velocities (See Appendix 1 for decriptions of the possible movement models). The effective sample sizes  generally decrease with less days in the data. The effective sample sizes were generally above 10 for sampling regimes that had at least six days (e.g. C6, S6), but dropped to worrying low quantities with only 3 days (e.g. C3, S3).

**Note on home range plots:** These plots show the 95% UD estimates with 95% CIs of the AKDE home range estimates for the 60 sampling regimes. 
Each plot shows the total locations from the complete, but the closed green points are the locations that were selected in the regimes, and the open pink points are the locations that were thinned out. 

The below figures and tables show the above information for each of the six groupings of sampling regimes (AA, AA2, RR, CE, SP, FL).

### I. AA Group


```{r, include=FALSE}
aa_fit <- readRDS("../Intermediate/ctmm/error/FITS_aa_error.rds") # read in model fits
```

```{r, include=FALSE, eval=FALSE}
aa_svf <- readRDS("Intermediate/ctmm/error/SVF_aa_error.rds") # read in variograms
# save image of variograms plotted with model fits
png(filename = "Figures/AA_SVF.png", width = 3000, height = 2200, units = "px")
plot_vario(aa_svf, aa_fit, 
           model_color = "#009E73", 
           cex = 3, fraction = 0.6, 
           relative_zoom = T,
           columns = 4,
           title_vec = names(aa_svf))
dev.off()
```


```{r, echo=FALSE, fig.cap="AA Variograms and Model Fits", fig.align='center', out.width = '95%'}
# read in image
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/AA_SVF.png")
```



```{r, include = FALSE, eval = FALSE}

### MODEL SUMMARY

# below not necessary anymore

## change names of sampling regimes (were originally named C1, C2, C3, etc.)
## changed to indicate number of days to make more informative
names(aa_fit) <- c("all","C30","C20","C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")

# change identity of fits
for (i in 1:length(names(aa_fit))){
  aa_fit[[i]]@info$identity <- names(aa_fit)[[i]]
}
```


```{r, echo = FALSE}

## function to get summaries of model fits
summarize_fits <- function(fit){
  
  summary <- summary(fit, units = FALSE) 
  
  tibble(individual.local.identifier = fit@info$identity,
         model = summary$name,
         hr_xing_time = (summary$CI[2,2]/3600), # convert seconds to hours
         effective_sample_size = (summary$DOF[2]))
  
}

# wrapper to stack info into data frame
make_summary_df <- function(id){
  map_dfr(id, summarize_fits) 
}

# apply function to get summary info
aa_fits_summaries <- make_summary_df(aa_fit)

# make table
knitr::kable(aa_fits_summaries,
             col.names = c("individual.local.identifier",
                           "Model Name",
                           "HR Crossing Time (hours)",
                           "Effective Sample Size"),
             caption = "Model summary information for AA group") 
```


```{r, include=FALSE, eval=FALSE}

## Home Range Plots

# Get boundaries and points and change to sf object

## first change everything to sf objects
## change names of sampling regimes (were originally named C1, C2, C3, etc.)
## changed to indicate number of days to make more informative

# change data to sf for mapping in ggplot
aa_data_sf <- readRDS("Intermediate/ctmm/error/DATA_aa_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% # map telemetery list into sf dataframe
  reduce(rbind) %>% 
  mutate(id = identity,
    # id = gsub("R", "S", identity),
    #      id = ifelse(id == "C1", "C30", 
    #                  ifelse(id == "C2", "C20",
    #                         ifelse(id == "C3", "C10",
    #                                ifelse(id == "C4", "C6",
    #                                       ifelse(id == "C5", "C3",
    #                                              ifelse(id == "S1", "S30",
    #                                                     ifelse(id == "S2", "S20",
    #                                                            ifelse(id == "S3", "S10",
    #                                                                   ifelse(id == "S4", "S6",
    #                                                                          ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  dplyr::select(id, geometry) %>% 
  filter(!id == "all") %>% # remove complete segment to map only sampling regimes
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0") # change to utm

# same procedure with HR mean boundary
aa_akde_sf <- readRDS("Intermediate/ctmm/error/AKDE_aa_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    # name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ), # get id by subsetting everything before space in name
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# again with HR confidence boundaries
aa_akde_CI <- readRDS("Intermediate/ctmm/error/AKDE_aa_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    # name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ),
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(!str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# do the same with locations from complete segment (to plot also the missing points)
aa_all_data <- readRDS("Intermediate/ctmm/error/DATA_aa_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity) %>% 
  #mutate(id = gsub("R", "S", identity)) %>% 
  dplyr::select(id, geometry) %>% 
  filter(id == "all")  %>% 
  dplyr::select(-id) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")
```


```{r, include=FALSE, eval=FALSE}

# Get data that was removed (missing locations) from complete segments to create sampling regimes. 
# These points will be shown on the maps to illustrate how well the HR estimates captured the missing data.
# this is done by doing anti_join of locations from the complete segment with the locations of each sampling regimes

## getting all thinned data
# get unique ids
ids <- unique(aa_data_sf$id)

# make list
aa_data_list <- aa_data_sf %>% 
  as.data.frame() %>% 
  group_split(id)

# change names of list
names(aa_data_list) <- ids

# make empty list to be filled
aa_thinned_data <- list()

# run loop over to get data that was thinned from each regime
for (i in ids){
  aa_thinned_data[[i]] <- anti_join(as.data.frame(aa_all_data), aa_data_list[[i]], by = "geometry")
}

# bind into data frame and make sf object
aa_thinned_data <- as.data.frame(do.call(rbind, aa_thinned_data)) %>%
  rownames_to_column("id") %>% 
  mutate(id = sub("\\..*", "", id)) %>% 
  st_as_sf(crs = "+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

```

```{r, include = FALSE, eval = FALSE}

## make plot

# get colors
colors <- c("missing points" = "#E1BE6A", "used points" = "#40B0A6")
shape <- c("missing points" = 1, "used points" = 16)
size <- c("missing points" = 0.8, "used points" = 1.2)
labels <- c("missing points", "used points")

# plot
p1 <- ggplot() +
  geom_sf(data = aa_data_sf, # data in sampling regimes
          aes(color = "used points", shape = "used points", size = "used points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = aa_thinned_data, # data in complete segment that is missing from sampling regimes
          aes(color = "missing points", shape = "missing points", size = "missing points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = aa_akde_CI, 
          inherit.aes = FALSE, 
          fill = NA, linetype = "dashed", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = aa_akde_sf,
          inherit.aes = FALSE, 
          fill =NA , alpha = 0.06, linewidth = 1) +
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
  coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"), 
           xlim=c(235.200,239.500), ylim=c(608.200, 611.400)) +
  scale_color_manual(values = colors, labels = labels, name = "") +
  scale_shape_manual(values = shape, labels = labels, name = "") +
  scale_size_manual(values = size, labels = labels, name = "") +
  scale_x_continuous(labels = scales::label_number(accuracy = .1)) +
  scale_y_continuous(labels = scales::label_number(accuracy = .1, big.mark = "")) +
  guides(color=guide_legend(override.aes=list(size = 8)))+
  labs( x = "Easting (km)", y = "Northing (km)") +
  theme_bw(base_family = "ArcherPro Book") +
  theme(legend.position="bottom", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        #axis.text.x = element_text(angle=45, hjust = 1),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 15)) +
  facet_wrap(~factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")), nrow = 2) 
```

```{r, include=FALSE, eval = FALSE}
# Save plot
file.name.performance <- paste0("AA_performance_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name.performance),
       plot = p1,
       width = 3400,
       height = 2000,
       units = "px")
```

```{r, echo=FALSE, fig.cap="Home Range Estimates from AA sampling regimes plotted over data", fig.align='center', out.width = '95%'}
# read in plot
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/AA_performance_plots_2023_09_13_122931.jpg")
```

### II. RR Group

```{r, include=FALSE}

### NOTE: SEE AA GROUP CODE FOR BETTER ANNOTATION -- THE SAME CODE IS USED FOR ALL GROUPS

rr_fit <- readRDS("../Intermediate/ctmm/error/FITS_rr_error.rds") # read in model fits
```

```{r, include=FALSE, eval=FALSE}
# plot variograms

rr_svf <- readRDS("Intermediate/ctmm/error/SVF_rr_error.rds") # read in variograms

png(filename = "Figures/RR_SVF.png", width = 3000, height = 2200, units = "px")
plot_vario(rr_svf, rr_fit, 
           model_color = "#56B4E9", 
           cex = 3, fraction = 0.6, 
           relative_zoom = T,
           columns = 4,
           title_vec = names(rr_svf))
dev.off()
```

```{r, echo=FALSE, fig.cap="RR Variograms and Model Fits", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/RR_SVF.png")
```

```{r, include = FALSE, eval = FALSE}
## change names of sampling regimes (were originally named C1, C2, C3, etc.)
## changed to indicate number of days to make more informative
# names(rr_fit) <- c("all","C30","C20","C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")
# 
# # change identity of fits
# for (i in 1:length(names(rr_fit))){
#   rr_fit[[i]]@info$identity <- names(rr_fit)[[i]]
# }
```


```{r, echo = FALSE}

# model summary table
rr_fits_summaries <- make_summary_df(rr_fit)

knitr::kable(rr_fits_summaries,
             col.names = c("individual.local.identifier",
                           "Model Name",
                           "HR Crossing Time (hours)",
                           "Effective Sample Size"),
             caption = "Model summary information for RR group")
```

```{r, include=FALSE, eval=FALSE}
# HR maps

# Get boundaries and points and change to sf object

## first change everything to sf objects

# data
rr_data_sf <- readRDS("Intermediate/ctmm/error/DATA_rr_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity,
    # id = gsub("R", "S", identity),
    #      id = ifelse(id == "C1", "C30", 
    #                  ifelse(id == "C2", "C20",
    #                         ifelse(id == "C3", "C10",
    #                                ifelse(id == "C4", "C6",
    #                                       ifelse(id == "C5", "C3",
    #                                              ifelse(id == "S1", "S30",
    #                                                     ifelse(id == "S2", "S20",
    #                                                            ifelse(id == "S3", "S10",
    #                                                                   ifelse(id == "S4", "S6",
    #                                                                          ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  dplyr::select(id, geometry) %>% 
  filter(!id == "all") %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR mean boundary
rr_akde_sf <- readRDS("Intermediate/ctmm/error/AKDE_rr_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ), # get id by subsetting everything before space in name
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR confidence boundaries
rr_akde_CI <- readRDS("Intermediate/ctmm/error/AKDE_rr_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ),
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(!str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# get all points
rr_all_data <- readRDS("Intermediate/ctmm/error/DATA_rr_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity) %>% 
  #mutate(id = gsub("R", "S", identity)) %>% 
  dplyr::select(id, geometry) %>% 
  filter(id == "all")  %>% 
  dplyr::select(-id) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")
```


```{r, include=FALSE, eval=FALSE}

# Get data that was removed from complete segments to create regimes. These points will be shown on the maps.

## getting all thinned data
# get unique ids
ids <- unique(rr_data_sf$id)

# make list
rr_data_list <- rr_data_sf %>% 
  as.data.frame() %>% 
  group_split(id)

# change names of list
names(rr_data_list) <- ids

# make empty list to be filled
rr_thinned_data <- list()

# run loop over to get data that was thinned from each regime
for (i in ids){
  rr_thinned_data[[i]] <- anti_join(as.data.frame(rr_all_data), rr_data_list[[i]], by = "geometry")
}

# bind into data frame and make sf object
rr_thinned_data <- as.data.frame(do.call(rbind, rr_thinned_data)) %>%
  rownames_to_column("id") %>% 
  mutate(id = sub("\\..*", "", id)) %>% 
  st_as_sf(crs = "+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

```

```{r, include = FALSE, eval = FALSE}

# plot
p2 <- ggplot() +
  geom_sf(data = rr_data_sf, # data in sampling regimes
          aes(color = "used points", shape = "used points", size = "used points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = rr_thinned_data, # data in complete segment that is missing from sampling regimes
          aes(color = "missing points", shape = "missing points", size = "missing points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = rr_akde_CI, 
          inherit.aes = FALSE, 
          fill = NA, linetype = "dashed", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = rr_akde_sf,
          inherit.aes = FALSE, 
          fill =NA , alpha = 0.06, linewidth = 1) +
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
  coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"), 
           xlim=c(235.800,239.200), ylim=c(607.200, 610.700)) +
  scale_color_manual(values = colors, labels = labels, name = "") +
  scale_shape_manual(values = shape, labels = labels, name = "") +
  scale_size_manual(values = size, labels = labels, name = "") +
  scale_x_continuous(labels = scales::label_number(accuracy = .1)) +
  scale_y_continuous(labels = scales::label_number(accuracy = .1, big.mark = "")) +
  guides(color=guide_legend(override.aes=list(size = 8)))+
  labs( x = "Easting (km)", y = "Northing (km)") +
  theme_bw(base_family = "ArcherPro Book") +
  theme(legend.position="bottom", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        #axis.text.x = element_text(angle=45, hjust = 1),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 15)) +
  facet_wrap(~factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")), nrow = 2) 
```

```{r, include=FALSE, eval = FALSE}
# Save plot
file.name.performance <- paste0("RR_performance_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name.performance),
       plot = p2,
       width = 3400,
       height = 2000,
       units = "px")
```

```{r, echo=FALSE, fig.cap="Home Range Estimates from RR sampling regimes plotted over data", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/RR_performance_plots_2023_09_13_123451.jpg")
```

### III. CE Group

```{r, include = FALSE}
ce_fit <- readRDS("../Intermediate/ctmm/error/FITS_ce_error.rds") # read in model fits
```

```{r, include = FALSE, eval = FALSE}
# plot variograms
ce_svf <- readRDS("Intermediate/ctmm/error/SVF_ce_error.rds") # read in variograms

png(filename = "Figures/CE_SVF.png", width = 3000, height = 2200, units = "px")
plot_vario(ce_svf, ce_fit, 
           model_color = "#E69F00", 
           cex = 3, fraction = 0.6, 
           relative_zoom = T,
           columns = 4,
           title_vec = names(ce_svf))
dev.off()

```

```{r, echo=FALSE, fig.cap="CE Variograms and Model Fits", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/CE_SVF.png")
```

```{r, include = FALSE, eval=FALSE}
## change names of sampling regimes (were originally named C1, C2, C3, etc.)
## changed to indicate number of days to make more informative
# names(ce_fit) <- c("all","C30","C20","C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")
# 
# # change identity of fits
# for (i in 1:length(names(ce_fit))){
#   ce_fit[[i]]@info$identity <- names(ce_fit)[[i]]
# }
```


```{r, echo = FALSE}
# plot model summary info
ce_fits_summaries <- make_summary_df(ce_fit)

knitr::kable(ce_fits_summaries,
             col.names = c("individual.local.identifier",
                           "Model Name",
                           "HR Crossing Time (hours)",
                           "Effective Sample Size"),
             caption = "Model summary information for CE group")
```


```{r, include = FALSE, eval = FALSE}

# HR plots

# Get boundaries and points and change to sf object
## change to sf objects

# data
ce_data_sf <- readRDS("Intermediate/ctmm/error/DATA_ce_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = factor(identity, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  dplyr::select(id, geometry) %>% 
  filter(!id == "all") %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR mean boundary
ce_akde_sf <- readRDS("Intermediate/ctmm/error/AKDE_ce_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ), # get id by subsetting everything before space in name
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR confidence boundaries
ce_akde_CI <- readRDS("Intermediate/ctmm/error/AKDE_ce_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ),
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>% 
  filter(!str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# get all points
ce_all_data <- readRDS("Intermediate/ctmm/error/DATA_ce_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>%
  mutate(id = identity) %>% 
  #mutate(id = gsub("R", "S", identity)) %>% 
  dplyr::select(id, geometry) %>% 
  filter(id == "all")  %>% 
  dplyr::select(-id) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")
```

```{r, include = FALSE, eval = FALSE}

# Get data that was removed from complete segments to create regimes. These points will be shown on the maps.

## getting all thinned data
# get unique ids
ids <- unique(ce_data_sf$id)

# make list
ce_data_list <- ce_data_sf %>% 
  #as.data.frame() %>% 
  group_split(id)

# change names of list
names(ce_data_list) <- ids

# make empty list to be filled
ce_thinned_data <- list()

# run loop over to get data that was thinned from each regime
for (i in ids){
  ce_thinned_data[[i]] <- anti_join(as.data.frame(ce_all_data), ce_data_list[[i]], by = "geometry")
}

# bind into data frame and make sf object
ce_thinned_data <- as.data.frame(do.call(rbind, ce_thinned_data)) %>%
  rownames_to_column("id") %>% 
  mutate(id = sub("\\..*", "", id)) %>% 
  st_as_sf(crs = "+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

```

```{r, include = FALSE, eval = FALSE}

size_ce <- c("missing points" = 1.5, "used points" = 1.2)

# plot
p3 <- ggplot() +
  geom_sf(data = ce_data_sf, # data in sampling regimes
          aes(color = "used points", shape = "used points", size = "used points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = ce_thinned_data, # data in complete segment that is missing from sampling regimes
          aes(color = "missing points", shape = "missing points", size = "missing points"),
          inherit.aes = FALSE, alpha = 1.2) +
  geom_sf(data = ce_akde_CI, 
          inherit.aes = FALSE, 
          fill = NA, linetype = "dashed", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = ce_akde_sf,
          inherit.aes = FALSE, 
          fill =NA , alpha = 0.06, linewidth = 1) +
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
  coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"), 
          xlim=c(234.500,236.500), ylim=c(609.400, 611.400)) +
  scale_color_manual(values = colors, labels = labels, name = "") +
  scale_shape_manual(values = shape, labels = labels, name = "") +
  scale_size_manual(values = size, labels = labels, name = "") +
  # scale_x_continuous(labels = scales::label_number(accuracy = .1)) +
  # scale_y_continuous(labels = scales::label_number(accuracy = .1, big.mark = "")) +
  guides(color=guide_legend(override.aes=list(size = 8)))+
  labs( x = "Easting (km)", y = "Northing (km)") +
  #theme_bw(base_family = "ArcherPro Book") +
  theme(legend.position="bottom", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        legend.key=element_blank(),
        #axis.text.x = element_text(angle=45, hjust = 1),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 15),
        panel.grid.major.y = element_blank(), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.y = element_blank(), 
        panel.grid.minor.x = element_blank(),
        panel.background = element_rect(fill = "white"),
        panel.border = element_rect(colour = "black", fill=NA, linewidth=0.5),
        strip.background = element_rect(color = "black", linewidth = 0.5)) +
  #guides(color = guide_legend(override.aes = list(size=5))) +
  facet_wrap(~factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")), nrow = 2)
```

```{r, include=FALSE, eval = FALSE}
# Save plot
file.name.performance <- paste0("CE_performance_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name.performance),
       plot = p3,
       width = 3400,
       height = 2000,
       units = "px")
```

```{r, echo=FALSE, fig.cap="Home Range Estimates from CE sampling regimes plotted over data", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/CE_performance_plots_2023_07_11_142209.jpg")
```

### IV. AA2 Group


```{r, include=FALSE}
aa2_fit <- readRDS("../Intermediate/ctmm/error/FITS_aa2_error.rds") # model fits
```

```{r, include=FALSE, eval=FALSE}
# plot variograms

aa2_svf <- readRDS("Intermediate/ctmm/error/SVF_aa2_error.rds") # read in variograms

png(filename = "Figures/AA2_SVF.png", width = 3000, height = 2200, units = "px")
plot_vario(aa2_svf, aa2_fit, 
           model_color = "#AA4499", 
           cex = 3, fraction = 0.6, 
           relative_zoom = T,
           columns = 4,
           title_vec = names(aa2_svf))
dev.off()

```

```{r, echo=FALSE, fig.cap="AA2 Variograms and Model Fits", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/AA2_SVF.png")
```


```{r, include = FALSE, eval = FALSE}
## change names of sampling regimes (were originally named C1, C2, C3, etc.)
## changed to indicate number of days to make more informative
# names(aa2_fit) <- c("all","C30","C20","C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")
# 
# # change identity of fits
# for (i in 1:length(names(aa2_fit))){
#   aa2_fit[[i]]@info$identity <- names(aa2_fit)[[i]]
# }
```


```{r, echo = FALSE}
# model summary table
aa2_fits_summaries <- make_summary_df(aa2_fit)

knitr::kable(aa2_fits_summaries,
             col.names = c("individual.local.identifier",
                           "Model Name",
                           "HR Crossing Time (hours)",
                           "Effective Sample Size"),
             caption = "Model summary information for AA2 group")
```


```{r, include=FALSE, eval=FALSE}

# make HR plots

# Get boundaries and points and change to sf object

## first change everything to sf objects

# data
aa2_data_sf <- readRDS("Intermediate/ctmm/error/DATA_aa2_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity,
    # id = gsub("R", "S", identity),
    #      id = ifelse(id == "C1", "C30", 
    #                  ifelse(id == "C2", "C20",
    #                         ifelse(id == "C3", "C10",
    #                                ifelse(id == "C4", "C6",
    #                                       ifelse(id == "C5", "C3",
    #                                              ifelse(id == "S1", "S30",
    #                                                     ifelse(id == "S2", "S20",
    #                                                            ifelse(id == "S3", "S10",
    #                                                                   ifelse(id == "S4", "S6",
    #                                                                          ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  dplyr::select(id, geometry) %>% 
  filter(!id == "all") %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR mean boundary
aa2_akde_sf <- readRDS("Intermediate/ctmm/error/AKDE_aa2_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ), # get id by subsetting everything before space in name
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR confidence boundaries
aa2_akde_CI <- readRDS("Intermediate/ctmm/error/AKDE_aa2_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ),
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(!str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# get all points
aa2_all_data <- readRDS("Intermediate/ctmm/error/DATA_aa2_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>%
  mutate(id = identity) %>% 
  #mutate(id = gsub("R", "S", identity)) %>% 
  dplyr::select(id, geometry) %>% 
  filter(id == "all")  %>% 
  dplyr::select(-id) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")
```


```{r, include=FALSE, eval=FALSE}

# Get data that was removed from complete segments to create regimes. These points will be shown on the maps.

## getting all thinned data
# get unique ids
ids <- unique(aa2_data_sf$id)

# make list
aa2_data_list <- aa2_data_sf %>% 
  as.data.frame() %>% 
  group_split(id)

# change names of list
names(aa2_data_list) <- ids

# make empty list to be filled
aa2_thinned_data <- list()

# run loop over to get data that was thinned from each regime
for (i in ids){
  aa2_thinned_data[[i]] <- anti_join(as.data.frame(aa2_all_data), aa2_data_list[[i]], by = "geometry")
}

# bind into data frame and make sf object
aa2_thinned_data <- as.data.frame(do.call(rbind, aa2_thinned_data)) %>%
  rownames_to_column("id") %>% 
  mutate(id = sub("\\..*", "", id)) %>% 
  st_as_sf(crs = "+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

```

```{r, include = FALSE, eval = FALSE}
# plot
p4 <- ggplot() +
  geom_sf(data = aa2_data_sf, # data in sampling regimes
          aes(color = "used points", shape = "used points", size = "used points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = aa2_thinned_data, # data in complete segment that is missing from sampling regimes
          aes(color = "missing points", shape = "missing points", size = "missing points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = aa2_akde_CI, 
          inherit.aes = FALSE, 
          fill = NA, linetype = "dashed", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = aa2_akde_sf,
          inherit.aes = FALSE, 
          fill =NA , alpha = 0.06, linewidth = 1) +
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
   coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"), 
           xlim=c(235.500,238.300), ylim=c(608.500, 611.200)) +
  scale_color_manual(values = colors, labels = labels, name = "") +
  scale_shape_manual(values = shape, labels = labels, name = "") +
  scale_size_manual(values = size, labels = labels, name = "") +
  scale_x_continuous(labels = scales::label_number(accuracy = .1)) +
  scale_y_continuous(labels = scales::label_number(accuracy = .1, big.mark = "")) +
  guides(color=guide_legend(override.aes=list(size = 8)))+
  labs( x = "Easting (km)", y = "Northing (km)") +
  theme_bw(base_family = "ArcherPro Book") +
  theme(legend.position="bottom", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        #axis.text.x = element_text(angle=45, hjust = 1),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 15)) +
  facet_wrap(~factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")), nrow = 2) 
```

```{r, include=FALSE, eval = FALSE}
# Save plot
file.name.performance <- paste0("AA2_performance_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name.performance),
       plot = p4,
       width = 3400,
       height = 2000,
       units = "px")
```

```{r, echo=FALSE, fig.cap="Home Range Estimates from AA2 sampling regimes plotted over data", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/AA2_performance_plots_2023_07_11_143200.jpg")
```

### V. SP Group

```{r, include=FALSE}
sp_fit <- readRDS("../Intermediate/ctmm/error/FITS_sp_error.rds") # read in model fits
```

```{r, include=FALSE, eval=FALSE}
# plot variograms
sp_svf <- readRDS("Intermediate/ctmm/error/SVF_sp_error.rds") # read in variograms

png(filename = "Figures/SP_SVF.png", width = 3000, height = 2200, units = "px")
plot_vario(sp_svf, sp_fit, 
           model_color = "#D55E00", 
           cex = 3, fraction = 0.6, 
           relative_zoom = T,
           columns = 4,
           title_vec = names(sp_svf))
dev.off()
```

```{r, echo=FALSE, fig.cap="SP Variograms and Model Fits", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/SP_SVF.png")
```

```{r, include = FALSE, eval = FALSE}
## change names of sampling regimes (were originally named C1, C2, C3, etc.)
## changed to indicate number of days to make more informative
# names(sp_fit) <- c("all","C30","C20","C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")
# 
# # change identity of fits
# for (i in 1:length(names(sp_fit))){
#   sp_fit[[i]]@info$identity <- names(sp_fit)[[i]]
# }
```


```{r, echo = FALSE}

# plot model summary info
sp_fits_summaries <- make_summary_df(sp_fit)

knitr::kable(sp_fits_summaries,
             col.names = c("individual.local.identifier",
                           "Model Name",
                           "HR Crossing Time (hours)",
                           "Effective Sample Size"),
             caption = "Model summary information for SP group")
```


```{r, include=FALSE, eval=FALSE}

# make HR plots

# Get boundaries and points and change to sf object

## first change everything to sf objects

# data
sp_data_sf <- readRDS("Intermediate/ctmm/error/DATA_sp_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity,
    # id = gsub("R", "S", identity),
    #      id = ifelse(id == "C1", "C30", 
    #                  ifelse(id == "C2", "C20",
    #                         ifelse(id == "C3", "C10",
    #                                ifelse(id == "C4", "C6",
    #                                       ifelse(id == "C5", "C3",
    #                                              ifelse(id == "S1", "S30",
    #                                                     ifelse(id == "S2", "S20",
    #                                                            ifelse(id == "S3", "S10",
    #                                                                   ifelse(id == "S4", "S6",
    #                                                                          ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  dplyr::select(id, geometry) %>% 
  filter(!id == "all") %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR mean boundary
sp_akde_sf <- readRDS("Intermediate/ctmm/error/AKDE_sp_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ), # get id by subsetting everything before space in name
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR confidence boundaries
sp_akde_CI <- readRDS("Intermediate/ctmm/error/AKDE_sp_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ),
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(!str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# get all points
sp_all_data <- readRDS("Intermediate/ctmm/error/DATA_sp_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity) %>% 
  #mutate(id = gsub("R", "S", identity)) %>% 
  dplyr::select(id, geometry) %>% 
  filter(id == "all")  %>% 
  dplyr::select(-id) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")
```


```{r, include=FALSE, eval=FALSE}

# Get data that was removed from complete segments to create regimes. These points will be shown on the maps.

## getting all thinned data
# get unique ids
ids <- unique(sp_data_sf$id)

# make list
sp_data_list <- sp_data_sf %>% 
  as.data.frame() %>% 
  group_split(id)

# change names of list
names(sp_data_list) <- ids

# make empty list to be filled
sp_thinned_data <- list()

# run loop over to get data that was thinned from each regime
for (i in ids){
  sp_thinned_data[[i]] <- anti_join(as.data.frame(sp_all_data), sp_data_list[[i]], by = "geometry")
}

# bind into data frame and make sf object
sp_thinned_data <- as.data.frame(do.call(rbind, sp_thinned_data)) %>%
  rownames_to_column("id") %>% 
  mutate(id = sub("\\..*", "", id)) %>% 
  st_as_sf(crs = "+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

```

```{r, include = FALSE, eval = FALSE}
# plot
p5 <- ggplot() +
  geom_sf(data = sp_data_sf, # data in sampling regimes
          aes(color = "used points", shape = "used points", size = "used points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = sp_thinned_data, # data in complete segment that is missing from sampling regimes
          aes(color = "missing points", shape = "missing points", size = "missing points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = sp_akde_CI, 
          inherit.aes = FALSE, 
          fill = NA, linetype = "dashed", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = sp_akde_sf,
          inherit.aes = FALSE, 
          fill =NA , alpha = 0.06, linewidth = 1) +
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
   coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"), 
           xlim=c(235.430,237.870), ylim=c(606.900, 609.550)) +
  scale_color_manual(values = colors, labels = labels, name = "") +
  scale_shape_manual(values = shape, labels = labels, name = "") +
  scale_size_manual(values = size, labels = labels, name = "") +
  scale_x_continuous(labels = scales::label_number(accuracy = .1)) +
  scale_y_continuous(labels = scales::label_number(accuracy = .1, big.mark = "")) +
  guides(color=guide_legend(override.aes=list(size = 8)))+
  labs( x = "Easting (km)", y = "Northing (km)") +
  theme_bw(base_family = "ArcherPro Book") +
  theme(legend.position="bottom", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        #axis.text.x = element_text(angle=45, hjust = 1),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 15)) +
  facet_wrap(~factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")), nrow = 2) 
```

```{r, include=FALSE, eval = FALSE}
# Save plot
file.name.performance <- paste0("SP_performance_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name.performance),
       plot = p5,
       width = 3400,
       height = 2000,
       units = "px")
```

```{r, echo=FALSE, fig.cap="Home Range Estimates from SP sampling regimes plotted over data", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/SP_performance_plots_2023_07_11_145830.jpg")
```

### VI. FL Group

```{r, include=FALSE}
fl_fit <- readRDS("../Intermediate/ctmm/error/FITS_fl_error.rds") # read in model fits
```

```{r, include=FALSE, eval=FALSE}
# plot variograms

fl_svf <- readRDS("Intermediate/ctmm/error/SVF_fl_error.rds") # read in variograms

png(filename = "Figures/FL_SVF.png", width = 3000, height = 2200, units = "px")
plot_vario(fl_svf, fl_fit, 
           model_color = "#997700", 
           cex = 3, fraction = 0.6, 
           relative_zoom = T,
           columns = 4,
           title_vec = names(fl_svf))
dev.off()
```

```{r, echo=FALSE, fig.cap="FL Variograms and Model Fits", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/FL_SVF.png")
```


```{r, include = FALSE, eval = FALSE}
## change names of sampling regimes (were originally named C1, C2, C3, etc.)
## changed to indicate number of days to make more informative
# names(fl_fit) <- c("all","C30","C20","C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")
# 
# # change identity of fits
# for (i in 1:length(names(fl_fit))){
#   fl_fit[[i]]@info$identity <- names(fl_fit)[[i]]
# }
```


```{r, echo = FALSE}
# model summary tables
fl_fits_summaries <- make_summary_df(fl_fit)

knitr::kable(fl_fits_summaries,
             col.names = c("individual.local.identifier",
                           "Model Name",
                           "HR Crossing Time (hours)",
                           "Effective Sample Size"),
             caption = "Model summary information for FL group")
```


```{r, include=FALSE, eval=FALSE}
#make HR plots

# Get boundaries and points and change to sf object

## first change everything to sf objects

# data
fl_data_sf <- readRDS("Intermediate/ctmm/error/DATA_fl_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity,
    # id = gsub("R", "S", identity),
    #      id = ifelse(id == "C1", "C30", 
    #                  ifelse(id == "C2", "C20",
    #                         ifelse(id == "C3", "C10",
    #                                ifelse(id == "C4", "C6",
    #                                       ifelse(id == "C5", "C3",
    #                                              ifelse(id == "S1", "S30",
    #                                                     ifelse(id == "S2", "S20",
    #                                                            ifelse(id == "S3", "S10",
    #                                                                   ifelse(id == "S4", "S6",
    #                                                                          ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  dplyr::select(id, geometry) %>% 
  filter(!id == "all") %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR mean boundary
fl_akde_sf <- readRDS("Intermediate/ctmm/error/AKDE_fl_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ), # get id by subsetting everything before space in name
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# HR confidence boundaries
fl_akde_CI <- readRDS("Intermediate/ctmm/error/AKDE_fl_error.rds") %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  mutate(
    #name = gsub("R", "S", name),
         id = gsub( " .*$", "", name ),
         # id = ifelse(id == "C1", "C30", 
         #             ifelse(id == "C2", "C20",
         #                    ifelse(id == "C3", "C10",
         #                           ifelse(id == "C4", "C6",
         #                                  ifelse(id == "C5", "C3",
         #                                         ifelse(id == "S1", "S30",
         #                                                ifelse(id == "S2", "S20",
         #                                                       ifelse(id == "S3", "S10",
         #                                                              ifelse(id == "S4", "S6",
         #                                                                     ifelse(id == "S5", "S3", id)))))))))),
         id = factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3"))) %>%
  filter(!str_detect(row.names(.), "est"),
         !id == "all") %>%
  dplyr::select(id, geometry) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# get all points
fl_all_data <- readRDS("Intermediate/ctmm/error/DATA_fl_error.rds") %>% 
  purrr::map(ctmm::as.sf) %>% 
  reduce(rbind) %>% 
  mutate(id = identity) %>% 
  #mutate(id = gsub("R", "S", identity)) %>% 
  dplyr::select(id, geometry) %>% 
  filter(id == "all")  %>% 
  dplyr::select(-id) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")
```


```{r, include=FALSE, eval=FALSE}

# Get data that was removed from complete segments to create regimes. These points will be shown on the maps.

## getting all thinned data
# get unique ids
ids <- unique(fl_data_sf$id)

# make list
fl_data_list <- fl_data_sf %>% 
  as.data.frame() %>% 
  group_split(id)

# change names of list
names(fl_data_list) <- ids

# make empty list to be filled
fl_thinned_data <- list()

# run loop over to get data that was thinned from each regime
for (i in ids){
  fl_thinned_data[[i]] <- anti_join(as.data.frame(fl_all_data), fl_data_list[[i]], by = "geometry")
}

# bind into data frame and make sf object
fl_thinned_data <- as.data.frame(do.call(rbind, fl_thinned_data)) %>%
  rownames_to_column("id") %>% 
  mutate(id = sub("\\..*", "", id)) %>% 
  st_as_sf(crs = "+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

```

```{r, include = FALSE, eval = FALSE}
# plot
p6 <- ggplot() +
  geom_sf(data = fl_data_sf, # data in sampling regimes
          aes(color = "used points", shape = "used points", size = "used points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = fl_thinned_data, # data in complete segment that is missing from sampling regimes
          aes(color = "missing points", shape = "missing points", size = "missing points"),
          inherit.aes = FALSE, alpha = 1) +
  geom_sf(data = fl_akde_CI, 
          inherit.aes = FALSE, 
          fill = NA, linetype = "dashed", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = fl_akde_sf,
          inherit.aes = FALSE, 
          fill =NA , alpha = 0.06, linewidth = 1) +
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
   coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"), 
           xlim=c(234.500,238.600), ylim=c(610.100, 614.100)) +
  scale_color_manual(values = colors, labels = labels, name = "") +
  scale_shape_manual(values = shape, labels = labels, name = "") +
  scale_size_manual(values = size, labels = labels, name = "") +
  scale_x_continuous(labels = scales::label_number(accuracy = .1)) +
  scale_y_continuous(labels = scales::label_number(accuracy = .1, big.mark = "")) +
  guides(color=guide_legend(override.aes=list(size = 8)))+
  labs( x = "Easting (km)", y = "Northing (km)") +
  theme_bw(base_family = "ArcherPro Book") +
  theme(legend.position="bottom", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        #axis.text.x = element_text(angle=45, hjust = 1),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 15)) +
  facet_wrap(~factor(id, levels = c("C30", "C20", "C10","C6", "C3", "S30", "S20", "S10", "S6", "S3")), nrow = 2) 
```

```{r, include=FALSE, eval = FALSE}
# Save plot
file.name.performance <- paste0("FL_performance_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name.performance),
       plot = p6,
       width = 3400,
       height = 2000,
       units = "px")
```

```{r, echo=FALSE, fig.cap="Home Range Estimates from FL sampling regimes plotted over data", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/FL_performance_plots_2023_07_11_150227.jpg")
```
