==> RMarkdowns/05_Statistical_Analysis.Rmd <==
---
title: 'Appendix 3: Varying Effects Plots and Model Posteriors'
date: \today
output:
  rmarkdown::pdf_document:
    latex_engine: xelatex
    extra_dependencies: ["float"]
    fig_caption: yes        
    includes:  
      in_header: preamble-latex.txt
header-includes:
  \floatplacement{figure}{H}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "H", out.extra = "")
```

## I. Background 

This supplementary document presents additional information related to our statistical analysis. Our aim was to model the performance of home range estimation, as defined in the Methods section, based on different characteristics of the sampling regimes. To achieve this, we employed three binomial Bayesian generalized linear mixed models.

The first model utilized a binary predictor variable to indicate whether the data in the sampling regimes were concentrated or spread, to predict performance. The second model used the number of locations as the predictor, while the third model used the number of unique weeks. We included an interaction between the predictors and the binary variable indicating spread or concentrated data in the last two models, and all models have varying slopes and intercepts per group.

In this document, we provide varying effects plots as referred to in the main text (but not presented there), along with the posterior distributions for all three statistical models.


```{r, eval = FALSE, include = FALSE}
# load packages
library(tidyverse)
library(bayesplot)
library(rstanarm)
library(magrittr)
library(rethinking)
library(janitor) # cleans up data for you
library(tidybayes)
library(tidybayes.rethinking)
library(modelr)
library(brms)
library(gghalves)
library(ggbeeswarm)
library(patchwork)
library(emmeans)
library(cmdstanr)
library(viridis)
library(broom.mixed)
library(ggthemes)
library(marginaleffects)
```

```{r, include = FALSE, eval =  FALSE}
# these options make stan run faster
rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())
```

```{r, include = FALSE, eval = FALSE}
# read in results from performance calculations
df <- readRDS("Intermediate/DF_Performance.rds") %>% 
  clean_names() %>% 
  rename(type = clumped_or_random,
         weeks = num_uniweeks,
         locations = num_points) %>% 
  mutate(group = factor(group),
         type = factor(ifelse(type == "S", "spread", "concentrated")),
         id = factor(id),
         weeks_std = (weeks - mean(weeks))/sd(weeks),
         locations_std = (locations - mean(locations))/sd(locations),
         xings_std = (xings - mean(xings))/sd(xings),
         hours = locations/2) %>% 
  dplyr::select(group,id,prop,freq,total_points,hr_area,weeks,locations, weeks_std,locations_std, xings, xings_std, type, hours, tau_low, tau, tau_high)
```


```{r, eval = FALSE, include = FALSE}
# Take a look at the data and get initial idea using box plots.
ggplot(data = df, aes(x = type, y = prop)) +
  geom_half_point(data = df, aes(color = type),
                  transformation = position_quasirandom(width = 0.1),
                  side = "l", size = 3, alpha = 0.5) +
  geom_half_boxplot(aes(fill = type), alpha = 0.7, side = "r", outlier.shape = NA) +
  scale_fill_viridis_d(option = "plasma", end = 0.8) +
  scale_color_viridis_d(option = "plasma", end = 0.8) +
  geom_hline(yintercept = 0.95, linetype = "dashed", color = "black", size = 1, alpha = 0.5) +
  scale_y_continuous(lim = c(0.2,1)) +
  guides(color = "none", fill = "none") +
  labs(y = "Performance Score") +
  theme_bw(base_family = "ArcherPro Book") +
  theme(axis.text = element_text(size = 15),
        axis.title.y = element_text(size = 20),
        axis.title.x = element_blank(),
        plot.title = element_text(size = 20))
```

## II. Model 1: HR Performance ~ Binary Predictor (spread vs. concentrated)

```{r, eval=FALSE, include = FALSE}

## Statistic Models ------------------------------

# We used binomial models with the number of locations that fall within the UD estimate indicating "successes", and the total points indicate the number of trials.

### Concentrated vs Spread Binary predictor

# set priors
priors_prop <- c(
  prior(normal(0,1.2), class = "b"),
  prior(normal(0,1.5), class = "Intercept"),
  prior(normal(0,1), class = "sd"),
  prior(lkj(3), class = "cor")
)

## model proportion performance 

# with group effects
m_type <- brm(freq | trials(total_points) ~ type + (type|group),
              data = df,
              family = "binomial",
              init  = "0",
              prior = priors_prop,
              
              control = list(adapt_delta = 0.99999,
                             max_treedepth = 14),
              chains = 4, iter = 1000, warmup = 500,
              cores = 4, seed = 520, 
              backend = "cmdstanr"
)

saveRDS(m_type, "Intermediate/Models/model_type.rds")
```


```{r, include=FALSE, eval=FALSE}
# read in model
m_type <- readRDS("Intermediate/Models/model_type.rds")

# Posterior validation check
pp_check(m_type, ndraws = 50)

# posteriors for model parameters
m1 <- mcmc_plot(m_type, type = "areas")
```

```{r, eval = FALSE, include = FALSE}

# Save mcmc plot
file.name <- paste0("m_type_mcmc_plot_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name),
       bg = "white",
       plot = m1,
       width = 3500,
       height = 3000,
       units = "px")

```


```{r, eval = FALSE, include = FALSE}
## Plot posterior predictions for group effect (varying slopes and intercepts)
# get posterior predictions
m_type_pred <- m_type %>% 
  predicted_draws(newdata = tibble(type = df$type,
                               total_points = df$total_points,
                               group = df$group)) %>% 
  mutate(.epred_prop = .prediction/total_points) # change to proportion
```


```{r, eval = FALSE, include = FALSE}
# plot posterior per group 
p1_groups <- ggplot(data = m_type_pred, aes(x = type, y = .epred_prop)) +
  geom_half_point(data = df, aes(color = type, y = prop),
                  transformation = position_quasirandom(width = 0.1),
                  side = "l", size = 3, alpha = 0.5) +
  stat_halfeye(aes(fill = type), alpha = 0.7, outlier.shape = NA) + 
  scale_fill_viridis_d(option = "plasma", end = 0.8) +
  scale_color_viridis_d(option = "plasma", end = 0.8) +
  geom_hline(yintercept = 0.95, linetype = "dashed", color = "black", size = 1, alpha = 0.5) +
  scale_y_continuous(lim = c(0.2,1)) +
  guides(color = "none", fill = "none") +
  labs(y = "Performance Score") +
  theme_bw() +
  theme(axis.text = element_text(size = 15),
        axis.title.y = element_text(size = 20),
        axis.title.x = element_blank(),
        plot.title = element_text(size = 20),
        strip.text = element_text(size = 20),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank()) +
  facet_wrap(~group)
```

```{r, eval = FALSE, include = FALSE}

# Save plot
file.name <- paste0("m_type_pred_groups",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".pdf")

ggsave(here::here("Figures/pdf", file.name),
       plot = p1_groups,
       width = 3500,
       height = 3000,
       units = "px")

```


```{r, echo=FALSE, fig.cap="Group effects (varying slopes and intercepts) for binomial model predicting home range estimation performance by a binary variable indicating whether the data in the sampling regimes are concentrated or spread." , fig.align='center' , out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/m_type_pred_groups2023_04_19_164443.jpg")
```


```{r, echo=FALSE, fig.cap="Posterior distributions for parameters in binomial model predicting home range estimation performance by a binary variable indicating whether the data in the sampling regimes are concentrated or spread. Group is included as a random effect (random slopes and intercepts). Posteriors include the median and 80\\% credible intervals.", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/m_type_mcmc_plot_2023_04_19_162753.jpg")
```


```{r, eval = FALSE, include = FALSE}

# get predictions
# get posterior predictions
m_type_pred <- m_type %>% 
  epred_draws(newdata = tibble(type = df$type,
                               total_points = df$total_points),
                  re_formula = NA) %>% 
  mutate(.epred_prop = .epred/total_points) # change to proportion

# Plot posterior across groups
p1 <- ggplot(data = m_type_pred, aes(x = type, y = .epred_prop)) +
  geom_half_point(data = df, aes(color = type, y = prop),
                  transformation = position_quasirandom(width = 0.1),
                  side = "l", size = 3, alpha = 0.5) +
  stat_pointinterval(aes(color = type), size = 10, alpha = 0.5) + 
  scale_fill_viridis_d(option = "plasma", end = 0.8) +
  scale_color_viridis_d(option = "plasma", end = 0.8) +
  geom_hline(yintercept = 0.95, linetype = "dashed", color = "black", size = 1, alpha = 0.5) +
  scale_y_continuous(lim = c(0.2,1)) +
  guides(color = "none", fill = "none") +
  labs(y = "Performance Score",
       title = "A") +
  theme_bw() +
  theme(axis.text = element_text(size = 15),
        axis.title.y = element_text(size = 20),
        axis.title.x = element_blank(),
        plot.title = element_text(size = 20),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())

```

```{r, eval = FALSE, include = FALSE}
# summary stats
point_interval(m_type_pred$.epred_prop[m_type_pred$type=="concentrated"])
point_interval(m_type_pred$.epred_prop[m_type_pred$type=="spread"])
```

## III. Model 2: HR Performance ~ Number of Locations 

```{r, eval=FALSE, include = FALSE}
### Number of locations as predictor --------------------
# set priors
reg_priors <- c(
  prior(normal(0,1), class = "b"),
  prior(normal(0,1.5), class = "Intercept"),
  prior(exponential(0.67), class = "sd"),
  prior(lkj(3), class = "cor")
)

# model with data
m_locs <- brm(
  bf(freq | trials(total_points) ~ locations_std + type + locations_std*type + (locations_std|group)),
  data = df,
  save_pars = save_pars(all=TRUE),
  family = binomial(),
  prior = reg_priors,
  init  = "0",
  control = list(adapt_delta = 0.99999,
                 max_treedepth = 13),
  chains = 4, iter = 2000, warmup = 1000,
  cores = 4, seed = 1234, 
  backend = "cmdstanr"
)

saveRDS(m_locs, "Intermediate/Models/model_locations.rds")
```

```{r, include = FALSE, eval = FALSE}

# read in model
m_locs <- readRDS("Intermediate/Models/model_locations.rds")

# posterior validation
pp_check(m_locs, ndraws = 50) 

# posterior distributions for model parameters
m2 <- mcmc_plot(m_locs, type = 'areas') 
```

```{r, eval = FALSE, include = FALSE}

# Save plot
file.name <- paste0("m_locs_mcmc_plot_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name),
       bg = "white",
       plot = m2,
       width = 3500,
       height = 3000,
       units = "px")

```


```{r, eval = FALSE, include = FALSE}

# Posterior predictions poulation (varying slopes and intercepts)
post_loc_draws <- df %>% 
  data_grid(locations_std = seq_range(-2:3, n =30),
            total_points = total_points,
            type = unique(type)) %>%
  #add_predicted_draws(m_locs, ndraws = 200) %>%
  add_epred_draws(m_locs, ndraws = 100, re_formula = NA) %>%
  mutate(.epred_prop = .epred/total_points)

# get mean prediction
mean_only <- post_loc_draws %>% 
  group_by(locations_std, type) %>%
  summarize(.epred_prop = mean(.epred_prop))

# get real number of locations for x axis sequence
num_locs <- sapply(X = (seq(-2, 3, by = 0.5)),
                   FUN =  function (x) {str_c( "(", as.character(round((x * sd(df$locations)) + mean(df$locations)), digits = 0), ")", sep = "")})

num_locs[2] <- "(0)"

```

```{r, include = FALSE, eval = FALSE}

# plot without group effects
p2 <- ggplot(data = post_loc_draws, 
             aes(x = locations_std, 
                 y = .epred_prop, 
                 color = type, 
                 fill = type)) +
  geom_line(stat = "smooth", # smooth posterior lines grouped by type
            aes(group = paste(type, .draw)), 
            alpha = 0.03, 
            se = FALSE, 
            lwd = 1.5) + 
  geom_line(data = mean_only, # add mean line
            stat = "smooth", 
            aes(group = type), 
            lwd = 1) +
  geom_point(data = df,
             aes(y = prop), show.legend = FALSE) +
  geom_hline(yintercept = 0.95, linetype = "dashed", color = "black", size = 1, alpha = 0.5) +
  # annotate("segment", x = 1.15, xend = 1.15, y = -Inf, yend = 0.95,
  #              linetype = "dashed",
  #              color = "#370499",
  #              size = 1,
  #              alpha = 0.5) +
  # annotate("segment", x = 0.9, xend = 0.9, y = -Inf, yend = 0.95,
  #              linetype = "dashed",
  #              color = "#feb82c",
  #              size = 1,
  #              alpha = 0.5) +
  scale_fill_viridis_d(option = "plasma", end = 0.8) +
  scale_color_viridis_d(option = "plasma", end = 0.8) + 
  scale_x_continuous(limits = c(-1.4,2.5), 
                     breaks = seq(-2, 3, by = 0.5),
                     labels = paste0(seq(-2, 3, by = 0.5), "\n", num_locs)) + # label real weeks underneath standardized
  labs(x = "Number of Recorded Locations",
       y = "Performance Score",
       title = "B") +
  theme_bw() +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 16),
        plot.title = element_text(size = 20),
        legend.title = element_blank(), 
        legend.position = "bottom",
        legend.key.size = unit(1, 'cm'), #change legend key size
        legend.key.height = unit(1, 'cm'), #change legend key height
        legend.key.width = unit(1, 'cm'), #change legend key width
        legend.text = element_text(size=13),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank()) +
  guides(color=guide_legend(title="Sampling Regularity")) 
  

p2
```

```{r, eval = FALSE, include = FALSE}

# Posterior predictions per group (varying slopes and intercepts)
post_loc_draws_grp <- df %>% 
  group_by(group) %>% 
  data_grid(locations_std = seq_range(-2:3, n =30),
            total_points = total_points,
            type = unique(type)) %>%
  #add_predicted_draws(m_locs, ndraws = 200) %>%
  add_predicted_draws(m_locs, ndraws = 100) %>%
  mutate(.epred_prop = .prediction/total_points)

# get mean predictions
mean_only_group <- post_loc_draws_grp %>% 
  group_by(locations_std, type, group) %>%
  summarize(.epred_prop = mean(.epred_prop))

# plot
p2_groups <- ggplot(data = post_loc_draws_grp, 
                    aes(x = locations_std, y = .epred_prop, color = type, fill = type)) +
  geom_line(stat = "smooth", 
            aes(group = paste(type, group, .draw)), 
            alpha = 0.05, se = FALSE, lwd = 1.5) + # smooth posterior lines grouped by type
  geom_line(data = mean_only_group, stat = "smooth", # add mean line
            aes(group = paste(type, group)), lwd = 1) +
  geom_point(data = df,
             aes(y = prop), show.legend = FALSE) +
  geom_hline(yintercept = 0.95, linetype = "dashed", color = "black", size = 1, alpha = 0.5) +
  scale_fill_viridis_d(option = "plasma", end = 0.8) +
  scale_color_viridis_d(option = "plasma", end = 0.8) + 
  scale_x_continuous(limits = c(-1.4,2.5), 
                     breaks = seq(-2, 3, by = 0.5),
                     labels = paste0(seq(-2, 3, by = 0.5), "\n", num_locs)) + # label real weeks underneath standardized
  labs(x = "Number of Recorded Locations",
       y = "Performance Score") +
  theme_clean() +
  theme_bw() +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 23),
        legend.title = element_blank(), 
        legend.position = "bottom",
        legend.key.size = unit(2,'cm'),
        legend.text = element_text(size=15),
        strip.text = element_text(size = 20),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank()) +
  facet_wrap(~group)
```

```{r, eval = FALSE, include = FALSE}

# Save plot
file.name <- paste0("m_locs_pred_groups",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".pdf")

ggsave(here::here("Figures/pdf", file.name),
       plot = p2_groups,
       width = 3800,
       height = 3000,
       units = "px")

```


```{r, echo=FALSE, fig.cap="Group effects (varying slopes and intercepts) for binomial model predicting home range estimation performance by number of locations as predictor with an interaction with the concentrated vs spread variable. Notice that FL and CE require the most locations to achive optimal performance, and are the least robust to low quantities of locations. These groups also have the most fragmented habitats (see Discussion section in the main text).", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/m_locs_pred_groups2023_04_28_163202.jpg")
```


```{r, echo=FALSE, fig.cap="Posterior distributions for parameters in binomial model predicting home range estimation performance by number of locations as predictor with an interaction with the concentrated vs spread variable. Group is included as a random effect (random slopes and intercepts). Posteriors include the median and 80\\% credible intervals.", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/m_locs_mcmc_plot_2023_04_19_163254.jpg")
```


## IV. Model 3: HR Performance ~ Number of Unique Weeks 

```{r, eval=FALSE, include = FALSE}
# model with data
m_weeks <- brm(
  bf(freq | trials(total_points) ~ weeks_std + type + weeks_std*type + (weeks_std|group)),
  data = df,
  save_pars = save_pars(all=TRUE),
  family = binomial(),
  prior = reg_priors,
  init  = "0",
  control = list(adapt_delta = 0.99999,
                 max_treedepth = 13),
  chains = 4, iter = 2000, warmup = 1000,
  cores = 4, seed = 1234, 
  backend = "cmdstanr"
)

saveRDS(m_weeks, "Intermediate/Models/model_weeks.rds")
```

```{r, eval = FALSE, include = FALSE}
# read in model
m_weeks <- readRDS("Intermediate/Models/model_weeks.rds")

# posterior validation
pp_check(m_weeks, ndraws = 50)

# get posteriors for model parameters
m3 <- mcmc_plot(m_weeks, type = 'areas') 
```

```{r, eval = FALSE, include = FALSE}

# Save plot
file.name <- paste0("m_weeks_mcmc_plot_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".jpg")

ggsave(here::here("Figures", file.name),
       bg = "white",
       plot = m3,
       width = 3500,
       height = 3000,
       units = "px")

```

```{r, include = FALSE, eval = FALSE}

# add posterior predictions
post_weeks_draws <- df %>% 
  data_grid(weeks_std = seq_range(-2:3, n =50),
            total_points = total_points,
            type = unique(type)) %>%
  add_epred_draws(m_weeks, ndraws = 100, re_formula = NA) %>% 
  mutate(.epred_prop = .epred/total_points)

# get mean prediction
mean_only <- post_weeks_draws %>% 
  group_by(weeks_std, type) %>%
  summarize(.epred_prop = mean(.epred_prop))

# get real number of weeks for x axis sequence
num_weeks <- sapply(X = (seq(-2, 3, by = 0.5)),
                    FUN =  function (x) {str_c( "(", as.character(round((x * sd(df$weeks)) + mean(df$weeks)), digits = 0), ")", sep = "")})

```


```{r, eval=FALSE, include = FALSE}

# plot without group effects
p3 <- ggplot(data = post_weeks_draws, 
       aes(x = weeks_std, 
           y = .epred_prop, 
           color = type, 
           fill = type)) +
  geom_line(stat = "smooth", # smooth posterior lines grouped by type
            aes(group = paste(type, .draw)), 
            alpha = 0.03, 
            se = FALSE, 
            lwd = 1.5) + 
  geom_line(data = mean_only, # add mean line
            stat = "smooth", 
            aes(group = type), 
            lwd = 1) +
  geom_point(data = df,
             aes(y = prop), show.legend = FALSE) +
  geom_hline(yintercept = 0.95, 
             linetype = "dashed",
             color = "black", 
             size = 1, alpha = 0.5) +
  # annotate("segment", x = 1.4, xend = 1.4, y = -Inf, yend = 0.95,
  #              linetype = "dashed",
  #              color = "#feb82c",
  #              size = 1,
  #              alpha = 0.5) +
  # annotate("segment", x = 0.35, xend = 0.35, y = -Inf, yend = 0.95,
  #              linetype = "dashed",
  #              color = "#370499",
  #              size = 1,
  #              alpha = 0.5) +
  scale_fill_viridis_d(option = "plasma", end = 0.8) +
  scale_color_viridis_d(option = "plasma", end = 0.8) + 
  scale_x_continuous(limits = c(-1.4,2.5), 
                     breaks = seq(-2, 3, by = 0.5),
                     labels = paste0(seq(-2, 3, by = 0.5), "\n", num_weeks)) + # label real weeks underneath standardized
  labs(x = "Number of Unique Weeks",
       y = "Performance Score",
       title = "C") +
  theme_bw() +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 16),
        plot.title = element_text(size = 20),
        legend.title = element_blank(), 
        legend.position = "bottom",
        legend.key.size = unit(1, 'cm'), #change legend key size
        legend.key.height = unit(1, 'cm'), #change legend key height
        legend.key.width = unit(1, 'cm'), #change legend key width
        legend.text = element_text(size=13),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank()) +
  guides(color=guide_legend(title="Sampling Regularity")) 

p3
```

```{r, include=FALSE, eval = FALSE}

p4 <- p1 + p2/p3 + plot_layout(guides = "collect") & theme(legend.position = 'bottom')

# Save plot
file.name.model.preds <- paste0("model_prediction_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".pdf")

ggsave(here::here("Figures/pdf", file.name.model.preds),
       plot = p4,
       width = 3500,
       height = 3000,
       units = "px")
```

```{r, include = FALSE, eval = FALSE}

# add posterior predictions per group
post_weeks_draws_grp <- df %>% 
  group_by(group) %>% 
  data_grid(weeks_std = seq_range(-2:3, n =50),
            total_points = total_points,
            type = unique(type)) %>%
  add_epred_draws(m_weeks, ndraws = 100) %>% 
  mutate(.epred_prop = .epred/total_points)

# get mean predictions
mean_only_group <- post_weeks_draws_grp %>% 
  group_by(weeks_std, type, group) %>%
  summarize(.epred_prop = mean(.epred_prop))

# plot group effects
p3_groups <- ggplot(data = post_weeks_draws_grp, 
                    aes(x = weeks_std, y = .epred_prop, color = type, fill = type)) +
  geom_line(stat = "smooth", aes(group = paste(type, group, .draw)), 
            alpha = 0.03, se = FALSE, lwd = 1.5) + # smooth posterior lines grouped by type
  geom_line(data = mean_only_group, stat = "smooth", # add mean line
            aes(group = paste(type,group)), lwd = 1) +
  geom_point(data = df,
             aes(y = prop), show.legend = FALSE) +
  geom_hline(yintercept = 0.95, linetype = "dashed", color = "black", size = 1, alpha = 0.5) +
  scale_fill_viridis_d(option = "plasma", end = 0.8) +
  scale_color_viridis_d(option = "plasma", end = 0.8) + 
  scale_x_continuous(limits = c(-1.4,2.5), 
                     breaks = seq(-2, 3, by = 0.5),
                     labels = paste0(seq(-2, 3, by = 0.5), "\n", num_weeks)) + # label real weeks underneath standardized
  labs(x = "Coarse-Scale Volume (# weeks)",
       y = "Performance Value") +
  theme_clean() +
  theme_bw() +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 23),
        legend.title = element_blank(), 
        legend.position = "bottom",
        legend.key.size = unit(2,'cm'),
        legend.text = element_text(size=15),
        strip.text = element_text(size = 20),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank()) +
  facet_wrap(~group)
```

```{r, eval = FALSE, include = FALSE}

# Save plot
file.name <- paste0("m_weeks_pred_groups",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".pdf")

ggsave(here::here("Figures/pdf", file.name),
       plot = p3_groups,
       width = 3800,
       height = 3000,
       units = "px")

```


```{r, echo=FALSE, fig.cap="Group effects (varying slopes and intercepts) for binomial model predicting home range estimation performance by number of unique weeks as predictor with an interaction with the concentrated vs spread variable. Notice that FL and CE require the most weeks to achive optimal performance, and are the least robust to low quantities of weeks. These groups also have the most fragmented habitats (see Discussion section in the main text).", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/m_weeks_pred_groups2023_04_28_163441.jpg")
```


```{r, echo=FALSE, fig.cap="Posterior distributions for parameters in binomial model predicting home range estimation performance by number of unique weeks as predictor with an interaction with the concentrated vs spread variable. Group is included as a random effect (random slopes and intercepts). Posteriors include the median and 80\\% credible intervals.", fig.align='center', out.width = '95%'}
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/m_weeks_mcmc_plot_2023_04_19_165345.jpg")
```


```{r, include = FALSE, eval = FALSE}

## Slope coefficient plots ---------------

# clumped locations
cred_intervals_locs_clumped <- m_locs %>% 
  emtrends(~ locations_std*type, var = "locations_std",
           #at = list(type = "concentrated", locations_std = c(-1.3, -0.8, -0.3, 0.2)),
           at = list(type = "concentrated", locations_std = c(-1.3, -0.55, 0.2)),
           regrid = "response") 

# plot table of intervals
cred_intervals_locs_clumped

# get osterior distributions
ame_locs_clumped <- cred_intervals_locs_clumped %>% 
  gather_emmeans_draws() %>% 
  mutate(amount = ifelse(locations_std == -1.3, "Low", 
                         ifelse(locations_std == -0.55, "Med", "High")),
         labels = str_c(amount,
                        " (",
                        as.character(round((locations_std * sd(df$locations)) + mean(df$locations)),
                                     digits = 0), 
                        " or ",
                        as.character(locations_std),
                        "SD",
                        ")", 
                        sep = ""))

#levels.1 <- c("-1.3 (21)", "-0.8 (142)", "-0.3 (264)", "0.2 (385)")
levels.1 <- c("Low (21 or -1.3SD)", "Med (203 or -0.55SD)", "High (385 or 0.2SD)")


p1 <- ggplot(ame_locs_clumped, aes(x = .value, 
                                   fill = factor(labels,levels = levels.1))) +
  geom_vline(xintercept = 0,
             linetype = "dashed",
             color = "black", 
             size = 1, alpha = 0.5) +
  stat_halfeye(.width = c(0.8, 0.95), point_interval = NULL,
               slab_alpha = 0.65) +
  stat_pointinterval(position = position_dodge(width = -1, preserve = "single"),
                     aes(x = .value, color = factor(labels, levels = rev(levels.1))),
                     alpha = 10, show.legend = FALSE) +
  scale_fill_viridis_d(option = "plasma", begin = 0, end = 0.4) +
  scale_color_viridis_d(option = "plasma", begin = 0.4, end = 0) +
  coord_cartesian(xlim = c(-0.05,0.5)) +
  labs(x = "Instantaneous Slope Coefficients", 
       y = "Density", fill = "Number of Locations",
       #caption = "80% and 95% credible intervals shown in black",
       title = "A",
       subtitle = "Concentrated Sampling") +
  theme_bw() +
  theme(legend.position=c(.75,.85),
        legend.title = element_text(size=14), #change legend title font size
        legend.text = element_text(size=11),
        legend.key.size = unit(0.5, 'cm'), #change legend key size
        legend.key.height = unit(0.5, 'cm'), #change legend key height
        legend.key.width = unit(0.5, 'cm'), #change legend key width 
        legend.background = element_rect(fill=alpha(0.1)),
        axis.title = element_blank(),
        axis.text = element_text(size = 12),
        plot.title = element_text(size = 20, vjust = -5),
        plot.subtitle = ggtext::element_markdown(hjust = 0.5, size = 15),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())

# spread locations
cred_intervals_locs_spread <- m_locs %>% 
  emtrends(~ locations_std*type, var = "locations_std",
           #at = list(type = "spread", locations_std = c(-1.3, -0.8, -0.3, 0.2)),
           at = list(type = "spread", locations_std = c(-1.3, -0.55, 0.2)),
           regrid = "response") 

# table of intervals
cred_intervals_locs_spread

# get posterior distributions
ame_locs_spread <- cred_intervals_locs_spread %>% 
  gather_emmeans_draws() %>% 
  mutate(amount = ifelse(locations_std == -1.3, "Low", 
                         ifelse(locations_std == -0.55, "Med", "High")),
         labels = str_c(amount,
                        " (",
                        as.character(round((locations_std * sd(df$locations)) + mean(df$locations)),
                                     digits = 0), 
                        " or ",
                        as.character(locations_std),
                        "SD",
                        ")", 
                        sep = ""))

p2 <- ggplot(ame_locs_spread, aes(x = .value, 
                                  fill = factor(labels,levels = levels.1))) +
  geom_vline(xintercept = 0,
             linetype = "dashed",
             color = "black", 
             size = 1, alpha = 0.5) +
  stat_halfeye(.width = c(0.8, 0.95), point_interval = NULL,
               slab_alpha = 0.65) +
  stat_pointinterval(position = position_dodge(width = -2, preserve = "single"),
                     aes(x = .value, color = factor(labels, levels = rev(levels.1))),
                     alpha = 10, show.legend = FALSE) +
  scale_fill_viridis_d(option = "plasma", begin = 0.6, end = 1) +
  scale_color_viridis_d(option = "plasma", begin = 1, end = 0.6) +
  coord_cartesian(xlim = c(-0.05,0.5)) +
  labs(x = "Instantaneous Slope Coefficients", 
       y = "Density", fill = "Number of Locations",
       #caption = "80% and 95% credible intervals shown in black",
       title = "B",
       subtitle = "Spread Sampling") +
  theme_bw() +
  theme(legend.position=c(.75,.85),
        legend.title = element_text(size=14), #change legend title font size
        legend.text = element_text(size=11),
        legend.key.size = unit(0.5, 'cm'), #change legend key size
        legend.key.height = unit(0.5, 'cm'), #change legend key height
        legend.key.width = unit(0.5, 'cm'), #change legend key width 
        legend.background = element_rect(fill=alpha(0.1)),
        axis.title = element_blank(),
        axis.text = element_text(size = 12),
        axis.text.y=element_blank(),
        plot.title = element_text(size = 20, vjust = -5),
        plot.subtitle = ggtext::element_markdown(hjust = 0.5, size = 15),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())

# clumped weekss
cred_intervals_weeks_clumped <- m_weeks %>% 
  emtrends(~ weeks_std*type, var = "weeks_std",
           #at = list(type = "concentrated", weeks_std = c(-1.3, -0.8, -0.3, 0.2)),
           at = list(type = "concentrated", weeks_std = c(-1.3, -0.55, 0.2)),
           regrid = "response") 

# table of intervals
cred_intervals_weeks_clumped

# get posterior distributions
ame_weeks_clumped <- cred_intervals_weeks_clumped %>% 
  gather_emmeans_draws() %>% 
  mutate(amount = ifelse(weeks_std == -1.3, "Low", 
                         ifelse(weeks_std == -0.55, "Med", "High")),
         labels = str_c(amount,
                        " (",
                        as.character(round((weeks_std * sd(df$weeks)) + mean(df$weeks)),
                                     digits = 0), 
                        " or ",
                        as.character(weeks_std),
                        "SD",
                        ")", 
                        sep = ""))

#levels.2 <- c("-1.3 (1)", "-0.8 (2)", "-0.3 (4)", "0.2 (6)")
levels.2 <- c("Low (1 or -1.3SD)", "Med (3 or -0.55SD)", "High (6 or 0.2SD)")

p3 <- ggplot(ame_weeks_clumped, aes(x = .value, 
                                    fill = factor(labels, levels = levels.2))) +
  geom_vline(xintercept = 0,
             linetype = "dashed",
             color = "black", 
             size = 1, alpha = 0.5) +
  stat_halfeye(.width = c(0.8, 0.95), point_interval = NULL,
               slab_alpha = 0.65) +
  stat_pointinterval(position = position_dodge(width = -1, preserve = "single"),
                     aes(x = .value, color = factor(labels, levels = rev(levels.2))),
                     alpha = 10, show.legend = FALSE) +
  scale_fill_viridis_d(option = "plasma", begin = 0, end = 0.4) +
  scale_color_viridis_d(option = "plasma", begin = 0.4, end = 0) +
  coord_cartesian(xlim = c(-0.05,0.5)) +
  labs(x = "Instantaneous Slope Coefficients", 
       y = "Density", fill = "Number of Weeks",
       title = "C") +
  theme_bw() +
  theme(legend.position=c(.75,.85),
        legend.title = element_text(size=14), #change legend title font size
        legend.text = element_text(size=11),
        legend.key.size = unit(0.5, 'cm'), #change legend key size
        legend.key.height = unit(0.5, 'cm'), #change legend key height
        legend.key.width = unit(0.5, 'cm'), #change legend key width 
        legend.background = element_rect(fill=alpha(0.1)),
        axis.text = element_text(size = 12),
        axis.title = element_blank(),
        plot.title = element_text(size = 20, vjust = -0.5),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())

# spread weeks
cred_intervals_weeks_spread <- m_weeks %>% 
  emtrends(~ weeks_std*type, var = "weeks_std",
           #at = list(type = "spread", weeks_std = c(-1.3, -0.8, -0.3, 0.2)),
           at = list(type = "spread", weeks_std = c(-1.3, -0.55, 0.2)),
           regrid = "response") 

# table of intervals
cred_intervals_weeks_spread

# get posterior distributions
ame_weeks_spread <- cred_intervals_weeks_spread %>% 
  gather_emmeans_draws() %>% 
   mutate(amount = ifelse(weeks_std == -1.3, "Low", 
                         ifelse(weeks_std == -0.55, "Med", "High")),
         labels = str_c(amount,
                        " (",
                        as.character(round((weeks_std * sd(df$weeks)) + mean(df$weeks)),
                                     digits = 0), 
                        " or ",
                        as.character(weeks_std),
                        "SD",
                        ")", 
                        sep = ""))

p4 <- ggplot(ame_weeks_spread, aes(x = .value, 
                                   fill = factor(labels, levels = levels.2))) +
  geom_vline(xintercept = 0,
             linetype = "dashed",
             color = "black", 
             size = 1, alpha = 0.5) +
  stat_halfeye(.width = c(0.8, 0.95), point_interval = NULL,
               slab_alpha = 0.65) +
  stat_pointinterval(position = position_dodge(width = -2, preserve = "single"),
                     aes(x = .value, color = factor(labels, levels = rev(levels.2))),
                     alpha = 10, show.legend = FALSE) +
  scale_fill_viridis_d(option = "plasma", begin = 0.6, end = 1) +
  scale_color_viridis_d(option = "plasma", begin = 1, end = 0.6) +
  coord_cartesian(xlim = c(-0.05,0.5)) +
  labs(x = "Instantaneous Slope Coefficients", 
       y = "Density", fill = "Number of Weeks",
       caption = "point intervals show 80% and 95% credible intervals",
       title = "D") +
  theme_bw() +
  theme(legend.position=c(.75,.85), # .7,.65
        legend.title = element_text(size=14), #change legend title font size
        legend.text = element_text(size=11),
        legend.key.size = unit(0.5, 'cm'), #change legend key size
        legend.key.height = unit(0.5, 'cm'), #change legend key height
        legend.key.width = unit(0.5, 'cm'), #change legend key width 
        legend.background = element_rect(fill=alpha(0.1)),
        axis.text = element_text(size = 12),
        axis.text.y=element_blank(),
        axis.title = element_blank(),
        plot.title = element_text(size = 20, vjust = -0.5),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())
```

```{r, include = FALSE, eval = FALSE}
## function to get global axes
add_global_label <- function(pwobj, Xlab = NULL, Ylab = NULL, Xgap = 0.03, Ygap = 0.03, ...) {
  ylabgrob <- patchwork::plot_spacer()
  if (!is.null(Ylab)) {
    ylabgrob <- ggplot() +
      geom_text(aes(x = .5, y = .5), label = Ylab, angle = 90, ...) +
      theme_void()
  }
  if (!is.null(Xlab)) {
    xlabgrob <- ggplot() +
      geom_text(aes(x = .5, y = .5), label = Xlab, ...) +
      theme_void()
  }
  if (!is.null(Ylab) & is.null(Xlab)) {
    return((ylabgrob + patchworkGrob(pwobj)) + 
             patchwork::plot_layout(widths = 100 * c(Ygap, 1 - Ygap)))
  }
  if (is.null(Ylab) & !is.null(Xlab)) {
    return((ylabgrob + pwobj) + 
             (xlabgrob) +
             patchwork::plot_layout(heights = 100 * c(1 - Xgap, Xgap),
                                    widths = c(0, 100),
                                    design = "
                                   AB
                                   CC
                                   "
             ))
  }
  if (!is.null(Ylab) & !is.null(Xlab)) {
    return((ylabgrob + pwobj) + 
             (xlabgrob) +
             patchwork::plot_layout(heights = 100 * c(1 - Xgap, Xgap),
                                    widths = 100 * c(Ygap, 1 - Ygap),
                                    design = "
                                   AB
                                   CC
                                   "
             ))
  }
  return(pwobj)
}

# plot together
p5 <- ((p1 + p2 + p3 + p4) + plot_layout(ncol = 2)) %>% 
  add_global_label(Ylab = "Density",
                   Xlab = "Instantaneous Slope Coefficients",
                   size = 7)

p5
```

```{r, include=FALSE, eval = FALSE}
# Save plot
file.name.model.slopes <- paste0("model_slope_plots_",
                                format(Sys.time(), "%Y_%m_%d_%H%M%S"), ".pdf")

ggsave(here::here("Figures/pdf", file.name.model.slopes),
       plot = p5,
       width = 3300,
       height = 3000,
       units = "px")
```
