==> RMarkdowns/07_Stationarity_Case_Study.Rmd <==
---
title: "Stationarity_Case_Study"
date: \today
output: 
  pdf_document:
    fig_caption: yes
    latex_engine: xelatex
    extra_dependencies: ["float"]
header-includes:
  \usepackage{fvextra}
  \DefineVerbatimEnvironment{Highlighting}{Verbatim}{breaklines,commandchars=\\\{\}}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.pos = "H", out.extra = "")
```

## Case Study: AA and RR group

In this Rmarkdown, we calculated a single annual home range for each group (2014 for AA and 2012 for RR) in which the data met our criteria for adequate home range estimation performance, as determined from our evaluation of the effects of sampling elements. Subsequently, we investigated changes in space-use over a period of four years by examining whether location data from different time periods, not included in the data used for home range estimation, fell within the boundaries of the home range. Specifically, we analyzed the proportion of locations from the previous year, the subsequent year, and the subsequent next year that fell within the mean 95% utilization distribution (UD) boundary of the home range estimate. This analysis allowed us to assess the consistency of the home ranges over time. Additionally, we visually assessed whether data from different seasons represented distinct portions of the annual home range. This approach enabled us to evaluate the temporal consistency of home ranges and determine the feasibility of segmenting the data based on specific time periods, such as years or seasons.

Load packages
```{r, message=FALSE, warning=FALSE, eval = FALSE}
# load packages
library(tidyverse)
library(ctmm)
library(ctmmweb)
library(sf)
library(sp)
library(lubridate)
library(mapview)
library(patchwork)
library(cowplot)
```

Read in data
```{r, eval= FALSE}
# read in raw data
df_raw <- read.csv("Data/CH1_GPS_data_case_study.csv", row.names = NULL)

# convert to telemetry object for ctmm
df_ctmm <- df_raw %>% 
  as.telemetry(projection = "+proj=utm +zone=10 +north +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")
```

Estimate Annual Home Ranges for AA 2014 and RR 2012
```{r, eval = FALSE}

# subset data
UD_data <- df_ctmm[c("aa_2014", "rr_2012")]

## FITTING UNCALIBRATED DATA UNDER A PRIOR

# assign 20-meter 2D RMS UERE
uere(UD_data) <- 20
UERE <- uere(UD_data)
UERE$DOF[] <- 2
summary(UERE)
uere(UD_data) <- UERE

## VARIOGRAMS, FITS, AKDES ........ 

SVF <- FITS <- AKDE <- list()

tictoc::tic()
for(i in 1:length(UD_data)){
  SVF[[i]] <- variogram(UD_data[[i]], 
                        dt = c(1,10) %#% "hour")
  GUESS <- ctmm.guess(UD_data[[i]], 
                      CTMM=ctmm(error=TRUE), 
                      interactive=FALSE, 
                      variogram = SVF[[i]])
  FITS[[i]] <- ctmm.select(UD_data[[i]],
                           GUESS,
                           trace=2, 
                           cores = 1)
  AKDE[[i]] <- akde(UD_data[[i]],
                    FITS[[i]], 
                    weights = TRUE, 
                    grid=list(dr=10,align.to.origin=TRUE))
}
tictoc::toc() 

names(AKDE) <- names(FITS)  <- names(SVF) <- names(UD_data)

# save AKDEs
saveRDS(AKDE, "Intermediate/ctmm/error/AKDE_case_study_error.rds")

```

Format data and UDs
```{r, eval = FALSE}
# read in AKDEs
UD_list <- readRDS("Intermediate/ctmm/error/AKDE_case_study_error.rds")
names(UD_list) <- c("aa_annual", "rr_annual")

# subset data from out of sample years
data_names <- c("aa_2013","aa_2015","aa_2016","rr_2011", "rr_2013" , "rr_2014" )
DATA_list <- df_ctmm[data_names]
```

Calculate proportion points within home range from previous year, subsequent year and subsequent next year

```{r, eval = FALSE}
# subset data out of sample by group
aa_dlist <- DATA_list[1:3]
names(aa_dlist) <- c("aa_prev_year", "aa_next_year", "aa_next_2year")

rr_dlist <- DATA_list[4:6]
names(rr_dlist) <- c("rr_prev_year", "rr_next_year", "rr_next_2year")

# Calculate proportion of points within HR for AA and RR
AAt <- AAp <-  AAp_low <- AAp_high <- AAf <-  AAf_low <- AAf_high <- list()
for(i in 1:length(aa_dlist)){
  AAt[[i]] <- SpatialPoints.telemetry(aa_dlist[i]) %over% SpatialPolygonsDataFrame.UD(UD_list[[1]] ,level.UD=0.95) %>% 
    table(useNA = "always") %>% # keep number of points that fell outside upper CI
    data.frame()  %>% 
    mutate(Prop = round(Freq/sum(Freq), digits = 3), # calculates proportion from frequency
           Freq = Freq); 
  AAf[[i]] <- AAt[[i]]$Freq[1] + AAt[[i]]$Freq[3]; # takes the proportion that fell within lower CI plus proportion that fell between mean and lower
  AAf_low[[i]] <- AAt[[i]]$Freq[3]; # proportion only within lower bound
  AAf_high[[i]] <- AAt[[i]]$Freq[1] + AAt[[i]]$Freq[2] + AAt[[i]]$Freq[3]; # all combined to get upper bound
  AAp[[i]] <- AAt[[i]]$Prop[1] + AAt[[i]]$Prop[3]; # takes the proportion that fell within lower CI plus proportion that fell between mean and lower
  AAp_low[[i]] <- AAt[[i]]$Prop[3]; # proportion only within lower bound
  AAp_high[[i]] <- AAt[[i]]$Prop[1] + AAt[[i]]$Prop[2] + AAt[[i]]$Prop[3]
}
names(AAt) <- names(AAp_low) <- names(AAp_high) <- names(AAp) <- names(AAf_low) <- names(AAf_high) <- names(AAf) <- names(aa_dlist)

# rr
RRt <- RRp <-  RRp_low <- RRp_high <- RRf <-  RRf_low <- RRf_high <- list()
for(i in 1:length(rr_dlist)){
  RRt[[i]] <- SpatialPoints.telemetry(rr_dlist[i]) %over% SpatialPolygonsDataFrame.UD(UD_list[[2]] ,level.UD=0.95) %>% 
    table(useNA = "always") %>% # keep number of points that fell outside upper CI
    data.frame()  %>% 
    mutate(Prop = round(Freq/sum(Freq), digits = 3), # calculates proportion from frequency
           Freq = Freq); 
  RRf[[i]] <- RRt[[i]]$Freq[1] + RRt[[i]]$Freq[3]; # takes the proportion that fell within lower CI plus proportion that fell between mean and lower
  RRf_low[[i]] <- RRt[[i]]$Freq[3]; # proportion only within lower bound
  RRf_high[[i]] <- RRt[[i]]$Freq[1] + RRt[[i]]$Freq[2] + RRt[[i]]$Freq[3]; # all combined to get upper bound
  RRp[[i]] <- RRt[[i]]$Prop[1] + RRt[[i]]$Prop[3]; # takes the proportion that fell within lower CI plus proportion that fell between mean and lower
  RRp_low[[i]] <- RRt[[i]]$Prop[3]; # proportion only within lower bound
  RRp_high[[i]] <- RRt[[i]]$Prop[1] + RRt[[i]]$Prop[2] + RRt[[i]]$Prop[3]
}
names(RRt) <- names(RRp_low) <- names(RRp_high) <- names(RRp) <- names(RRf_low) <- names(RRf_high) <- names(RRf) <- names(rr_dlist)
```

Organize into dataframe
```{r, eval = FALSE}
# put in dataframe
props_df <- tibble(ID = names(aa_dlist),
                   Group = "AA",
                   Prop = as.numeric(unlist(AAp)),
                   Prop_low = as.numeric(unlist(AAp_low)),
                   Prop_high = as.numeric(unlist(AAp_high)),
                   Freq = as.numeric(unlist(AAf)),
                   Freq_low = as.numeric(unlist(AAf_low)),
                   Freq_high = as.numeric(unlist(AAf_high))) %>%
  rbind(tibble(ID = names(rr_dlist),
               Group = "RR",
               Prop = as.numeric(unlist(RRp)),
               Prop_low = as.numeric(unlist(RRp_low)),
               Prop_high = as.numeric(unlist(RRp_high)),
               Freq = as.numeric(unlist(RRf)),
               Freq_low = as.numeric(unlist(RRf_low)),
               Freq_high = as.numeric(unlist(RRf_high)))) 
```

Format home ranges and data to sf for ggplot

```{r, eval = FALSE}
### convert to sf objects for plotting
# transform mean boundries to sf
mean_hr_sf <- UD_list %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  filter(str_detect(row.names(.), "est")) %>% 
  rename(id = name) %>%
  mutate(id = gsub( " .*$", "", id ),
         group = str_sub (id, 1,2)) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# transform confidence intervals to sf
ci_hr_sf <- UD_list %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  filter(!str_detect(row.names(.), "est")) %>% 
  rename(id = name) %>%
  mutate(id = gsub( " .*$", "", id ),
         group = str_sub (id, 1,2)) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# transform data to sf
data_sf <- DATA_list %>% 
  purrr::map(ctmm::as.sf, level.UD = 0.95) %>% 
  reduce(rbind) %>% 
  rename(id = identity) %>%
  mutate(group = str_sub (id, 1,2),
         season = ifelse(month(timestamp) %in% c(1:6), "Dry Season", "Wet Season")) %>% 
  st_transform("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0")

# separate groups and add scale column
aa_data_sf <- data_sf %>%
  filter(group == "aa") %>% 
  mutate(scale = ifelse(year(timestamp) == 2013,
                        "AA Data Previous Year (2013)",
                        ifelse(year(timestamp) == 2015,
                               "AA Data Following Year (2015)",
                               "AA Data 2 Years After (2016)")),
         prop_mean = ifelse(scale == "AA Data Previous Year (2013)",
                            props_df$Prop[props_df$ID=="aa_prev_year"],
                            ifelse(scale == "AA Data Following Year (2015)",
                                   props_df$Prop[props_df$ID=="aa_next_year"],
                                   props_df$Prop[props_df$ID=="aa_next_2year"])),
         prop = str_c( "Prop = ", as.character(prop_mean)))

rr_data_sf <- data_sf %>%
  filter(group == "rr") %>% 
  mutate(scale = ifelse(year(timestamp) == 2011,
                        "RR Data Previous Year (2011)",
                        ifelse(year(timestamp) == 2013,
                               "RR Data Following Year (2013)",
                               "RR Data 2 Years After (2014)")),
         prop_mean = ifelse(scale == "RR Data Previous Year (2011)",
                            props_df$Prop[props_df$ID=="rr_prev_year"],
                            ifelse(scale == "RR Data Following Year (2013)",
                                   props_df$Prop[props_df$ID=="rr_next_year"],
                                   props_df$Prop[props_df$ID=="rr_next_2year"])),
         prop = str_c( "Prop = ", as.character(prop_mean)))
```

Make Plots 
```{r, eval = FALSE}
colors <- c("Wet Season" = "#0072b2", "Dry Season" = "#d55e00")

# ggplot
p1 <- ggplot() +
  geom_sf(data = aa_data_sf, 
          aes(color = season),
          inherit.aes = FALSE, alpha = 1.5, shape = 1) +
  scale_color_manual(values = colors, labels = labels, name = "") +
  geom_sf(data = ci_hr_sf[ci_hr_sf$id=="aa_2014",], # HR CIs
          inherit.aes = FALSE,
          fill = NA, linetype = "dotted", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = mean_hr_sf[mean_hr_sf$id=="aa_2014",], # HR mean boundary
          inherit.aes = FALSE,
          fill =NA , alpha = 0.06, linewidth = 0.8) +
  geom_text(data = aa_data_sf, aes(x = 239.0, y = 608.4, label = prop), size = 4) +
  ggspatial::annotation_scale(location = "bl",
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
  coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"),
           xlim=c(235.200,239.500), ylim=c(608.400, 611.400)) +
  labs( x = "Easting (km)", y = "Northing (km)") +
  theme(legend.position="none", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        legend.key=element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 12),
        panel.grid.major.y = element_blank(), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.y = element_blank(), 
        panel.grid.minor.x = element_blank(),
        panel.background = element_rect(fill = "white"),
        panel.border = element_rect(colour = "black", fill=NA, linewidth=0.5),
        strip.background = element_rect(color = "black", linewidth = 0.5)) +
  guides(color = guide_legend(override.aes = list(size=5))) +
  facet_wrap(~factor(scale, levels = c("AA Data Previous Year (2013)", "AA Data Following Year (2015)", "AA Data 2 Years After (2016)")))

p2 <- ggplot() +
  geom_sf(data = rr_data_sf, 
          aes(color = season),
          inherit.aes = FALSE, alpha = 1.5, shape = 1) +
  scale_color_manual(values = colors, name = "") +
  geom_sf(data = ci_hr_sf[ci_hr_sf$id=="rr_2012",], # HR CIs
          inherit.aes = FALSE,
          fill = NA, linetype = "dotted", alpha = 0.8, linewidth = 0.2) +
  geom_sf(data = mean_hr_sf[mean_hr_sf$id=="rr_2012",], # HR mean boundary
          inherit.aes = FALSE,
          fill =NA , alpha = 0.06, linewidth = 0.8) +
  geom_text(data = rr_data_sf, aes(x = 238.8, y = 607.400, label = prop), size = 4) +
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.3,
                              height = unit(4,'pt'),
                              style = 'ticks') +
  coord_sf(datum = st_crs("+proj=utm +zone=10 +north +datum=WGS84 +units=km +no_defs +ellps=WGS84 +towgs84=0,0,0"),
           xlim=c(235.200,239.500), ylim=c(607.400, 610.250)) +
  labs( x = "Easting (km)", y = "Northing (km)") +
  theme(legend.position="bottom", 
        legend.title = element_blank(),
        legend.text=element_text(size=14),
        legend.key=element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        strip.text = element_text(size = 12),
        panel.grid.major.y = element_blank(), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.y = element_blank(), 
        panel.grid.minor.x = element_blank(),
        panel.background = element_rect(fill = "white"),
        panel.border = element_rect(colour = "black", fill=NA, linewidth=0.5),
        strip.background = element_rect(color = "black", linewidth = 0.5)) +
  guides(color = guide_legend(override.aes = list(size=5))) +
  facet_wrap(~factor(scale, levels = c("RR Data Previous Year (2011)", "RR Data Following Year (2013)", "RR Data 2 Years After (2014)")))

p3 <- egg::ggarrange(p1,p2)
```

```{r, echo=FALSE, fig.cap="Annual home range estimates for two groups of white-faced capuchins at the Lomas Barbudal Monkey Project in Costa Rica (2011-2016) plotted alongside data from the preceding, subsequent, and next subsequent year. Dark lines indicate the mean 95\\% utilization distribution contour for 2014 (group AA) and 2012 (group RR). Dotted lines indicate the corresponding 95\\% confidence intervals.", fig.align='center', out.width = '95%'}
# read in plot
knitr::include_graphics("C:/Users/ojacobson/Documents/PhD/R/CH1_Sampling/Figures/stationarity_plots2023_06_30_131330.jpg")
```
